Add case for 0 integrations found in node installer & tidy up
